requirements.txt
llama-index
python-dotenv
Pillow
PyMuPDF
openai
nest-asyncio
scikit-learn
torch
torchvision
tiktoken